Feature/fix bugs

+++ src/main/java/com/PascuanSilvestre/TorqTrace/features/workOrder/workOrder/dto/WorkOrderCreateDTO.java
@@ -24,7 +24,7 @@ public class WorkOrderCreateDTO {
 
     @NotNull
     @Min(0)
-    private Double entryKm;
+    private Integer entryKm;
 
 
     @NotBlank(message = "Description is required")
